api/requirements.txt
quart>=0.19.5,<0.20.0
pytest>=8.2.0,<9.0.0
hypercorn>=0.14.0,<0.15.0
pymongo>=4.7.2,<5.0.0
azure-storage-queue>=12.10.0,<13.0.0
azure-cosmos>=4.6.0,<5.0.0
python-dotenv>=1.0.1,<2.0.0
azure-storage-blob>=12.20.0,<13.0.0
quart-cors>=0.7.0,<0.8.0
